more validations and success messages. relations

+++ anser/app/Http/Controllers/ListCategoryController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use App\ListCategory;
+use App\ListCategory, App\Point;
 
 class ListCategoryController extends Controller
 {
@@ -36,14 +36,23 @@ class ListCategoryController extends Controller
      */
     public function store(Request $request)
     {
+
+        $messages = [
+            'category.required' => 'Kategorian nimi ei voi olla tyhjä',
+            'category.unique' => 'Kategoria on jo olemassa',
+            'category.max' => 'Liian pitkä nimi'
+        ];
+
         $validatedData = $request->validate([
         'category' => 'required|unique:list_categories|max:100'
-        ]);
+        ], $messages
+        );
 
         $listcategory = new ListCategory;
         $listcategory->category = request('category');
         $listcategory->save();
 
+        $request->session()->flash('alert-success', 'Kategoria tallennettu');
         $listcategorys = ListCategory::all();
         return view('listcategorys', compact('listcategorys'));
     }
@@ -88,10 +97,16 @@ class ListCategoryController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(ListCategory $listcategory)
+    public function destroy(Request $request, ListCategory $listcategory)
     {
+        $used = Point::UsedListCategories();
 
-        $listcategory->delete();
+        if ( $used -> contains( $listcategory->id ) === false) {
+            $listcategory->delete();
+            $request->session()->flash('alert-success', 'Kategoria poistettu');
+        } else {
+            $request->session()->flash('alert-danger', 'Kategoriaa ei poistettu, koska on olemassa siihen kuuluvia pinnatietoja');
+        }
         return redirect ('listcategorys');
 
          // return public function (
